Implement String.Split with count
JavaScript's split method has a different treatment of the limit
parameter than the CLR, and so these overloads had not been implemented.
This commit retrofits the JavaScript version with the correct semantics
on single-delimiter splits, and raises an error if multiple delimiters
are provided with a count. The test case is updated to check this
behaviour.

Partially fixes #646.